# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Insider SMS API",
        "description": "Otomatik mesaj gönderim sistemi API dokümantasyonu",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Development Server"
        }
    ],
    "paths": {
        "/api/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Gönderilmiş mesajları listele",
                "operationId": "4216f3eedb7be911361f1b91030253a8",
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "Test mesajı"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "sent"
                                                    },
                                                    "external_message_id": {
                                                        "type": "string",
                                                        "example": "msg_123456"
                                                    },
                                                    "sent_at": {
                                                        "type": "string",
                                                        "format": "datetime"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "datetime"
                                                    },
                                                    "recipients": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "phone_number": {
                                                                    "type": "string",
                                                                    "example": "+905551234567"
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "John Doe"
                                                                },
                                                                "status": {
                                                                    "type": "string",
                                                                    "example": "sent"
                                                                },
                                                                "external_message_id": {
                                                                    "type": "string",
                                                                    "example": "msg_123456"
                                                                },
                                                                "sent_at": {
                                                                    "type": "string",
                                                                    "format": "datetime"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Yeni mesaj oluştur",
                "operationId": "c85a9171473a13f151a8672235e01559",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content",
                                    "recipients"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string",
                                        "maxLength": 160,
                                        "example": "Test mesajı"
                                    },
                                    "recipients": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "phone_number": {
                                                    "type": "string",
                                                    "example": "+905551234567"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mesaj başarıyla oluşturuldu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Mesaj başarıyla oluşturuldu."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation hatası"
                    }
                }
            }
        },
        "/api/messages/{id}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Belirli bir mesajı getir",
                "operationId": "2d72cf14bf2d55c9c9dfd297da04d404",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Mesaj ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Mesaj bulunamadı"
                    }
                }
            }
        },
        "/api/messages/sent/list": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Gönderilmiş mesajların ID listesini getir",
                "operationId": "ae5cf3a1a00af22da188791ebc37f648",
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "message_ids": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "msg_123456"
                                                    }
                                                },
                                                "count": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/messages/status/{messageId}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Mesaj durumunu kontrol et (Redis cache'den)",
                "operationId": "7824d86d257086a11788f28b7f12fedb",
                "parameters": [
                    {
                        "name": "messageId",
                        "in": "path",
                        "description": "External Message ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Messages",
            "description": "Mesaj yönetimi API endpoints"
        }
    ]
}